+++ src/Effect.hs
@@ -1,7 +1,9 @@
 module Effect
     ( module Effect.Persistence
+    , module Effect.Interactive
     , module Effect.Error
     ) where
 
 import Effect.Error
+import Effect.Interactive
 import Effect.Persistence